글쓰기 DB 저장 완료

+++ src/main/java/com/its/board/dto/BoardDTO.java
@@ -27,5 +27,5 @@ public class BoardDTO {
     //
     private String storedFileName;
     //
-    private int member_id;
+    private Long member_id;
 }
